CLAN6-158 Refactor board config validator
