+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/config/validation/Validator.kt
@@ -1,0 +1,8 @@
+package com.cognifide.cogboard.config.validation
+
+import io.vertx.core.json.JsonObject
+
+interface Validator {
+
+    fun validate(config: JsonObject): Boolean
+}
