+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/config/validation/boards/BoardsValidationError.kt
@@ -1,0 +1,5 @@
+package com.cognifide.cogboard.config.validation.boards
+
+import com.cognifide.cogboard.config.model.Board
+
+data class BoardsValidationError(val board: Board, val message: String)
