+++ cogboard-app/src/main/kotlin/com/cognifide/cogboard/config/validation/boards/BoardsValidationException.kt
@@ -1,0 +1,3 @@
+package com.cognifide.cogboard.config.validation.boards
+
+class BoardsValidationException(message: String?) : Exception(message)
